requirements.txt
setuptools~=50.3.0
python-telegram-bot~=13.0